Move futures_unordered() into futures_unordered module
