+++ futures-util/src/stream/mod.rs
@@ -92,7 +92,7 @@ if_std! {
     pub use self::collect::Collect;
     pub use self::select_all::SelectAll;
     pub use self::split::{SplitStream, SplitSink, ReuniteError};
-    pub use self::futures_unordered::FuturesUnordered;
+    pub use self::futures_unordered::{futures_unordered, FuturesUnordered};
     pub use self::futures_ordered::{futures_ordered, FuturesOrdered};
 }
 
@@ -960,30 +960,6 @@ pub trait StreamExt: Stream {
     {
         recover::new(self, f)
     }
-}
-
-/// Converts a list of futures into a `Stream` of results from the futures.
-///
-/// This function will take an list of futures (e.g. a vector, an iterator,
-/// etc), and return a stream. The stream will yield items as they become
-/// available on the futures internally, in the order that they become
-/// available. This function is similar to `buffer_unordered` in that it may
-/// return items in a different order than in the list specified.
-///
-/// Note that the returned set can also be used to dynamically push more
-/// futures into the set as they become available.
-#[cfg(feature = "std")]
-pub fn futures_unordered<I>(futures: I) -> FuturesUnordered<<I::Item as IntoFuture>::Future>
-    where I: IntoIterator,
-        I::Item: IntoFuture
-{
-    let mut set = FuturesUnordered::new();
-
-    for future in futures {
-        set.push(future.into_future());
-    }
-
-    return set
 }
 
 /// Convert a list of streams into a `Stream` of results from the streams.
